+++ library/src/components/filter/FilterProperty.vue
@@ -9,6 +9,14 @@
         :class="{'selected': value === option.value, 'custom-selected': value === 'custom', 'join-selected': value === 'join' }"
         @click="update(option.value)"
       ) {{ option.label }}
+        NioTooltip(
+          v-if="option.tooltip"
+          :message="option.tooltip"
+          nudge-top
+          open-on-hover
+          content-class="nio-filter-option-tooltip"
+          :open-delay="400"
+        )
         .option-selected-pointer(v-if="(value === 'custom' && option.value === 'custom') || (value === 'join' && option.value === 'join')")
     .custom-option(
       v-if="value === 'custom'"
@@ -29,6 +37,7 @@
 <script>
 
 import NioFilterJoinOption from './join/JoinOption'
+import NioTooltip from '../Tooltip'
 
 export default {
   name: 'nio-filter-property',
@@ -64,9 +73,17 @@ export default {
       this.$emit('update:value', value)
     }
   },
-  components: { NioFilterJoinOption }
+  components: { NioFilterJoinOption, NioTooltip }
 }
 </script>
+
+<style lang="sass">
+  .nio-filter-option-tooltip
+    transform: translate(-7px, -90px) !important
+    width: 272px !important
+    min-width: 272px !important
+    // transform: translateX(-90px) !important
+</style>
 
 <style lang="sass" scoped>
   @import '@/styles/mixins/filter/_filter-property'