+++ library/src/modules/app/theme/theme.js
@@ -8,6 +8,8 @@ function setTheme() {
 }
 
 function getThemeColor(colorName) {
+	console.log(GlobalColors)
+	console.log(colorName)
 	let cssVar = document.documentElement.style.getPropertyValue(`--nio-theme-${toKebabCase(colorName)}`)
 	return cssVar ? cssVar : GlobalColors[colorName].default
 }