+++ CHANGELOG.md
@@ -114,3 +114,4 @@ v0.4.2
 Added support for Array attributes in **attributeModule**, **NioSchema** and **NioSchemaProperties**
 Added Jasmine for unit testing and added test for **attributeModule**
 Added default filter type for Array property type
+Added tooltip to display property type in **NioSchema** and **NioSchemaProperties**