+++ CHANGELOG.md
@@ -116,3 +116,4 @@ Added Jasmine for unit testing and added test for **attributeModule**
 Added default filter type for Array property type
 Added tooltip to display property type in **NioSchema** and **NioSchemaProperties**
 Added optional tooltip to filter option in **NioFilterProperty**
+Added **getJoinOptionsByPath** to **attributeModule** exports