+++ CHANGELOG.md
@@ -119,3 +119,6 @@ Added optional tooltip to filter option in **NioFilterProperty**
 Added **getJoinOptionsByPath** to **attributeModule** exports
 Added **nioItemIndex** to item click handler for **NioSlatTable** items template to allow tracking items with complex ids
 Added **indeterminate** loading option to **NioFileChooser**
+
+v0.4.3
+Added **NioPrettySchemaPath** component